pyproject.toml
[project]
name = "screenshotmacro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "imagehash>=4.3.2",
    "keyboard>=0.13.5",
    "pillow>=11.1.0",
    "pre-commit>=4.1.0",
    "pyautogui>=0.9.54",
    "pynput>=1.7.7",
    "typer[all]>=0.9.0",
]

[project.scripts]
screenshot-macro = "src.cli:main"

[tool.black]
target-version = ['py313']
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''
preview = true

[tool.isort]
py_version = 313
line_length = 100
atomic = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = []
known_third_party = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
pythonpath = ["."]
